Update README.md
How do you ascertain the security of a crypto coin when looking out for a long term investment, do you consider the vesting period or the multi signature access. The $LMTdc coin is a dynamic approach to future investment , building it security upon the smartest blockchain technology. 

Join here👇👇and secure your investment.
https://lmtdc.com/?utm_medium=social&utm_source=Github&utm_term=lmt&utm_content=Esio